streamablehttp/pyproject.toml
[project]
name = "streamablehttp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "anthropic>=0.76.0",
    "arxiv>=2.4.0",
    "fastmcp>=2.14.4",
    "google-cloud-storage>=3.8.0",
    "mcp>=1.26.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
